WebGL related part GraphicsDevice is moving to a separate class
